+++ src/mock/data-sort.js
@@ -1,0 +1,9 @@
+const sortData = [
+  { id: 'day', name: 'Day', checked: true, disabled: false },
+  { id: 'event', name: 'Event', checked: false, disabled: true },
+  { id: 'time', name: 'Time', checked: false, disabled: false },
+  { id: 'price', name: 'Price', checked: false, disabled: false },
+  { id: 'offer', name: 'Offers', checked: false, disabled: true },
+];
+
+export { sortData };
