+++ src/view/trip-filters-view.js
